cleanup and split to increase readability
