--- example/botd.json
@@ -1,35 +1,0 @@
-{
-  "apiVersion": "v1",
-  "kind": "ReplicationController",
-  "metadata": {
-    "name": "botdrc",
-    "labels": {
-      "name": "botdrc"
-    }
-  },
-  "spec": {
-    "replicas": 1,
-    "template": {
-      "metadata": {
-        "name": "botdrc",
-        "labels": {
-          "name": "botrc"
-        }
-      },
-      "spec": {
-        "containers": [
-          {
-            "name": "botd",
-            "image": "ipedrazas/botd:latest",
-						"ports":[
-                     {
-                        "containerPort":9090,
-                        "protocol":"TCP"
-                     }
-                  ]
-          }
-        ]
-      }
-    }
-  }
-}
