+++ yaml2json.go
@@ -1,0 +1,44 @@
+package main
+
+import (
+	"log"
+	"strconv"
+)
+
+func transformData(in interface{}) (out interface{}, err error) {
+	switch in.(type) {
+	case map[interface{}]interface{}:
+		o := make(map[string]interface{})
+		for k, v := range in.(map[interface{}]interface{}) {
+			sk := ""
+			switch k.(type) {
+			case string:
+				sk = k.(string)
+			case int:
+				sk = strconv.Itoa(k.(int))
+			default:
+				log.Fatal(err)
+			}
+			v, err = transformData(v)
+			if err != nil {
+				return nil, err
+			}
+			o[sk] = v
+		}
+		return o, nil
+	case []interface{}:
+		in1 := in.([]interface{})
+		len1 := len(in1)
+		o := make([]interface{}, len1)
+		for i := 0; i < len1; i++ {
+			o[i], err = transformData(in1[i])
+			if err != nil {
+				return nil, err
+			}
+		}
+		return o, nil
+	default:
+		return in, nil
+	}
+	return in, nil
+}
